+++ internal/producer/src/accountstable/accountstable.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accounts/accountinfo"
+	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accountinfo"
 )
 
 /*
