+++ internal/producer/src/block/block.go
@@ -10,7 +10,7 @@ import (
 	"reflect"
 	"time"
 
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accounts/contracts"
+	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
 )
